DDPG Actor model implementation using Keras
Note that the raw actions produced by the output layer are in a
[0.0, 1.0] range (using a sigmoid activation function).
So, we add another layer that scales each output to the desired range
for each action dimension. This produces a deterministic action for
any given state vector. A noise will be added later to this action to
produce some exploratory behavior.
Another thing to note is how the loss function is defined using
action value (Q value) gradients:
These gradients will need to be computed using the critic model,
and fed in while training. Hence it is specified as part of
the "inputs" used in the training function.